Update: Adding pull request suggested changes.

+++ parser.rb
@@ -1,27 +1,24 @@
 class ParserGamesLogs
-  attr_accessor :header_line 
-
   def initialize(filename_with_path)
     @filename_with_path = filename_with_path
     @mode = 'r'
-    @header_line = read_first_line()
-    
-    unless is_header_valid?()
-      raise "Invalid file - The header does not match a valid pattern"
-    end
+    file_exists?()
   end
   
-  private
 
   def read_first_line()
-    File.open(@filename_with_path, @mode) do |log_file|
-      log_file.readline
+    if file_exists?
+      File.open(@filename_with_path, @mode) do |log_file|
+        log_file.readline
+      end
     end
   end
 
+
+  private
   
-  def is_header_valid?()
-    header_pattern = "0:00 ------------------------------------------------------------"
-    return @header_line.strip == header_pattern
+  def file_exists?()
+    raise Errno::ENOENT unless File.exist?(@filename_with_path)
+    return true
   end
 end
